+++ lib/src/page_request.dart
@@ -1,19 +1,22 @@
 import 'pageable.dart';
 import 'sort.dart';
 
-class PageRequest implements Pageable {
+/// Abstract implementation of [Pageable].
+abstract class AbstractPageRequest implements Pageable {
   @override
   final int pageNumber;
 
   @override
   final int pageSize;
 
-  @override
-  final Sort sort;
-
-  const PageRequest({required this.pageNumber, required this.pageSize, this.sort = Sort.unsorted})
-    : assert(pageNumber >= 0, "Page index must not be less than zero!"),
-      assert(pageSize >= 1, "Page size must not be less than one!");
+  /// Creates a new [AbstractPageRequest]. Pages are zero indexed, thus providing 0 for [pageNumber] will
+  /// return the first page.
+  ///
+  /// [pageNumber] must not be negative.
+  /// [pageSize] must be greater than 0.
+  const AbstractPageRequest({required this.pageNumber, required this.pageSize})
+    : assert(pageNumber >= 0, 'Page index must not be less than zero'),
+      assert(pageSize >= 1, 'Page size must not be less than one');
 
   @override
   bool get isPaged => true;
@@ -25,21 +28,77 @@ class PageRequest implements Pageable {
   int get offset => pageNumber * pageSize;
 
   @override
+  bool hasPrevious() => pageNumber > 0;
+
+  @override
+  Pageable previousOrFirst() => hasPrevious() ? previous() : first();
+
+  @override
+  Pageable next();
+
+  /// Returns the [Pageable] requesting the previous page.
+  Pageable previous();
+
+  @override
+  Pageable first();
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(this, other)) return true;
+
+    return other is AbstractPageRequest &&
+        other.pageNumber == pageNumber &&
+        other.pageSize == pageSize;
+  }
+
+  @override
+  int get hashCode => Object.hash(pageNumber, pageSize);
+}
+
+/// Basic implementation of [Pageable].
+class PageRequest extends AbstractPageRequest {
+  @override
+  final Sort sort;
+
+  /// Creates a new [PageRequest] with sort parameters applied.
+  ///
+  /// [pageNumber] zero-based page number, must not be negative.
+  /// [pageSize] the size of the page to be returned, must be greater than 0.
+  /// [sort] sorting parameters, defaults to [Sort.unsorted].
+  const PageRequest({super.pageNumber = 0, required super.pageSize, this.sort = Sort.unsorted});
+
+  @override
   PageRequest next() => PageRequest(pageNumber: pageNumber + 1, pageSize: pageSize, sort: sort);
 
+  @override
   PageRequest previous() => pageNumber == 0
       ? this
       : PageRequest(pageNumber: pageNumber - 1, pageSize: pageSize, sort: sort);
 
   @override
-  PageRequest previousOrFirst() => hasPrevious() ? previous() : first();
-
-  @override
   PageRequest first() => PageRequest(pageNumber: 0, pageSize: pageSize, sort: sort);
 
+  /// Creates a new [PageRequest] with [pageNumber] applied.
   @override
-  Pageable withPage(int page) => PageRequest(pageNumber: page, pageSize: pageSize, sort: sort);
+  PageRequest withPage(int pageNumber) =>
+      PageRequest(pageNumber: pageNumber, pageSize: pageSize, sort: sort);
+
+  /// Creates a new [PageRequest] with [Sort] applied.
+  PageRequest withSort(Sort sort) =>
+      PageRequest(pageNumber: pageNumber, pageSize: pageSize, sort: sort);
 
   @override
-  bool hasPrevious() => pageNumber > 0;
+  bool operator ==(Object other) {
+    if (identical(this, other)) return true;
+
+    return other is PageRequest && super == other && other.sort == sort;
+  }
+
+  @override
+  int get hashCode => Object.hash(super.hashCode, sort);
+
+  @override
+  String toString() {
+    return 'Page request [number: $pageNumber, size: $pageSize, sort: $sort]';
+  }
 }